client/intelligent_collab/services/effects_service/pyproject.toml
[tool.poetry]
name = "effects container"
version = "0.1.0"
description = "effects ai service"
authors = ["Your Name <you@example.com>"]
license = "BSD"

# intel_tensorflow = "2.4.*"

[tool.poetry.dependencies]
python = "3.8.*"
fastapi = "^0.63.0"
typer = "^0.3.2"
uvicorn = "^0.13.4"
tensorflow-hub = "^0.12.0"
plotbbox = "^0.1.3"
msgpack = "^1.0.2"
msgpack-numpy = "^0.4.7"
Pillow = "^8.2.0"
matplotlib = "^3.4.2"
requests = "^2.25.1"
loguru = "^0.5.3"
tf-slim = "^1.1.0"
psutil = "^5.8.0"
pycocotools = "^2.0.2"
contextlib2 = "^0.6.0"
Cython = "^0.29.23"
pyzmq = "^22.1.0"
scikit-image = "^0.18.1"

[tool.poetry.dev-dependencies]
black = "^21.4b2"
pytest = "^6.2.3"
pytest-black = "^0.3.12"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
